+++ flake.nix
@@ -30,7 +30,7 @@
       ];
 
       pname = "tmux";
-      version = "0.1.1";
+      version = "0.2.0";
 
       homeModule = import ./modules {
         inherit pname;