+++ CHANGELOG.md
@@ -2,12 +2,19 @@
 
 # Changelog
 
-## 0.1.2
+## 0.2.0
+
+### Added
+
+- [line-numbers](https://www.reddit.com/r/tmux/comments/1kozf64/copy_mode_with_relative_line_numbers/) script
+- [tmux-fingers](https://github.com/Morantron/tmux-fingers) plugin
 
 ### Changed
 
 - changed status-left-length from 11 to 24
 - truncate session name at 16 chars
+- unbind time key `t`
+- update prompt cursor, message, and border colors
 
 ## 0.1.1